+++ protected/humhub/modules/ui/filter/widgets/views/textInput.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ *
+ */
+
+use yii\bootstrap\Html;
+
+/* @var $this \humhub\components\View */
+?>
+<div class="form-group">
+    <?= Html::textInput(null,null, $options) ?>
+</div>