+++ static/js/humhub/humhub.ui.js
@@ -1,0 +1,8 @@
+/**
+ * Contains user interface related modules and utilities.
+ *
+ * @namespace humhub.modules.ui
+ * @property {module:ui/status} status
+ **/
+
+// Only for documentation