+++ protected/humhub/modules/ui/form/widgets/DatePicker.php
@@ -9,8 +9,12 @@
 namespace humhub\modules\ui\form\widgets;
 
 use Yii;
+use yii\helpers\FormatConverter;
+use yii\helpers\Json;
 use yii\jui\DatePicker as BaseDatePicker;
 use humhub\libs\Html;
+use yii\jui\DatePickerLanguageAsset;
+use yii\jui\JuiAsset;
 
 /**
  * DatePicker form field widget
@@ -21,6 +25,21 @@ use humhub\libs\Html;
  */
 class DatePicker extends BaseDatePicker
 {
+    const LANGUAGEMAPPING = [
+        'en_gb' => 'en-GB',
+        'pt_br' => 'pt-BR',
+        'nb_no' => 'nb',
+        'nn_no' => 'nn',
+        'zh_cn' => 'zh-CN',
+        'zh_tw' => 'zh_TW',
+        'fa_ir' => 'fa',
+        'an' => null,
+        'uz' => null,
+        'ht' => null,
+        'am' => null,
+    ];
+
+    public $pickerLanguage;
 
     /**
      * @inheritdoc
@@ -34,6 +53,62 @@ class DatePicker extends BaseDatePicker
         Html::addCssClass($this->options, 'form-control');
 
         parent::init();
+
+        $this->pickerLanguage = $this->language ? $this->language : Yii::$app->language;
+        $this->pickerLanguage = (array_key_exists($this->pickerLanguage, static::LANGUAGEMAPPING))
+            ? static::LANGUAGEMAPPING[$this->pickerLanguage]
+            : $this->pickerLanguage;
+
+        if(!$this->pickerLanguage) {
+            $this->pickerLanguage = 'en-US';
+        }
+
+        $this->pickerLanguage = str_replace('_', '-', $this->pickerLanguage);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function run()
+    {
+        /**
+         * HUMHUB PATCH: Language Mapping + Prevent loading language files for all english based languages, since DatePickerLanguageAsset tries
+         * to load a fallback language e.g. for `en-GB` -> 'en'.
+         *
+         *
+         */
+
+        echo $this->renderWidget() . "\n";
+
+        $containerID = $this->inline ? $this->containerOptions['id'] : $this->options['id'];
+        $language = $this->language ? $this->language : Yii::$app->language;
+
+        if (strncmp($this->dateFormat, 'php:', 4) === 0) {
+            $this->clientOptions['dateFormat'] = FormatConverter::convertDatePhpToJui(substr($this->dateFormat, 4));
+        } else {
+            $this->clientOptions['dateFormat'] = FormatConverter::convertDateIcuToJui($this->dateFormat, 'date', $language);
+        }
+
+        if ($this->pickerLanguage !== 'en-US' && $this->pickerLanguage !== 'en') {
+            $view = $this->getView();
+
+            $assetBundle = DatePickerLanguageAsset::register($view);
+            if(substr($this->pickerLanguage, 0 , 2) === 'en') {
+                $assetBundle->autoGenerate = false;
+                $assetBundle->js[] = "ui/i18n/datepicker-$this->pickerLanguage.js";
+            } else {
+                $assetBundle->language = $this->pickerLanguage;
+            }
+
+            $options = Json::htmlEncode($this->clientOptions);
+            $this->pickerLanguage = Html::encode($this->pickerLanguage);
+            $view->registerJs("jQuery('#{$containerID}').datepicker($.extend({}, $.datepicker.regional['{$this->pickerLanguage}'], $options));");
+        } else {
+            $this->registerClientOptions('datepicker', $containerID);
+        }
+
+        $this->registerClientEvents('datepicker', $containerID);
+        JuiAsset::register($this->getView());
     }
 
 }