+++ protected/humhub/docs/CHANGELOG.md
@@ -18,6 +18,7 @@ HumHub Change Log
 - Fix: Comment form files not cleared
 - Enh: Added `humhub\modules\file\models\isAssignedTo($record)` argument in order to check if a File is attached to a specific record
 - Fix: rich-text mobile view wrong min-height calculation
+- Fix #3314: layout container width differences
 
 
 1.3.4  (September 25, 2018)